Better Defaults
In response to https://github.com/fjrdomingues/autopilot/issues/132
I think this is a better starting point for now. More info in the ticket.

We got 4k
We want the biggest file size to be able to edit
So 2k for the response.
That means the biggest file we can edit <2k
And Summaries < 2k, IDK how we get 800 tokens extra, but 🤷 should be no issue with 1k summaries I think.